requirements.txt
python-telegram-bot>=20.0
google-genai>=1.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pytz>=2024.1

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
